Moving ETag to ::Azure
