+++ sdk/core/azure-core/test/ut/match_conditions.cpp
@@ -1,12 +1,13 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // SPDX-License-Identifier: MIT
 
-#include <azure/core/etag.hpp>
 #include <azure/core/match_conditions.hpp>
+#include <azure/etag.hpp>
 #include <gtest/gtest.h>
 #include <string>
 
 using namespace Azure::Core;
+using namespace Azure;
 
 TEST(MatchConditions, Basic)
 {
