+++ sdk/core/azure-core/test/ut/simplified_header.cpp
@@ -17,6 +17,7 @@
 #pragma clang diagnostic ignored "-Wunused-variable"
 #endif // _MSC_VER
 
+#include <azure.hpp>
 #include <azure/core.hpp>
 #include <gtest/gtest.h>
 
@@ -32,7 +33,7 @@ TEST(SimplifiedHeader, core)
   EXPECT_NO_THROW(Azure::Core::CaseInsensitiveSet iset);
   EXPECT_NO_THROW(Azure::Core::Context c);
   EXPECT_NO_THROW(Azure::Core::DateTime(2020, 11, 03, 15, 30, 44));
-  EXPECT_NO_THROW(Azure::Core::ETag e);
+  EXPECT_NO_THROW(Azure::ETag e);
   EXPECT_NO_THROW(Azure::Core::Base64Decode("foo"));
   EXPECT_NO_THROW(Azure::Core::Cryptography::Md5Hash m);
   EXPECT_NO_THROW(Azure::Core::Http::RawResponse r(
